[feat] distance 단위 변환

+++ modules/application/src/main/kotlin/com/moyeobus/application/route/model/RouteDistanceRanking.kt
@@ -3,5 +3,5 @@ package com.moyeobus.application.route.model
 data class RouteDistanceRanking (
     val routeNo: Long,
     val ranking: Int,
-    val distance: Int
+    val distance: Double
 )
